backend/requirements.txt
# Django Core
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3

# Authentication & Security
djangorestframework-simplejwt==5.3.1
setuptools<81.0.0  # Pin setuptools to avoid pkg_resources deprecation warning
                   # NOTE: We also have a monkey patch in apps.authentication.patches.jwt_version
                   # that replaces pkg_resources with importlib.metadata for this library
django-oauth-toolkit==1.7.1
cryptography==42.0.2
argon2-cffi==23.1.0

# Database
psycopg[binary]==3.2.9

# Celery & Task Queue
celery==5.3.4
redis==5.0.1
kombu==5.3.4
flower==2.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.0

# API Documentation
drf-spectacular==0.27.1

# Data & Serialization
django-jsonfield==1.4.1
jsonschema==4.21.1
marshmallow==3.20.2

# Image Processing
Pillow==10.4.0

# HTTP and External APIs
requests==2.31.0
urllib3==2.1.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
python-decouple==3.8
pytz==2023.4
python-dateutil==2.9.0

# Health Checks & Monitoring
django-health-check==3.17.0
django-redis==5.4.0
django-ratelimit==4.1.0

# Development & Debugging
ipython==8.20.0
ipdb==0.13.13
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
flake8==6.1.0
black==23.12.1
isort==5.13.2

# WSGI/ASGI Server
gunicorn==21.2.0
uvicorn==0.27.0
daphne==4.0.0

# Environment & Configuration
environs==11.0.0
dj-database-url==2.1.0

# WebSocket Support
channels==4.0.0
channels-redis==4.2.0

# Data Processing
pandas==2.2.0
numpy==1.26.3